requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pinecone-client>=2.2.4
numpy>=1.24.0
python-dotenv>=1.0.0
